+++ VersionHistory.md
@@ -12,6 +12,7 @@ Prefix the description of the change with `[major]`, `[minor]` or `[patch]` in a
 * [major] Remove `JsonPointer` and `JsonPatch`.
 * [major] Simplify `JsonInputSettings` and `JsonOutputSettings`.
 * [major] Rename `JsonFilter.AlternatePathSepartor` to `AlternatePathSeparator`.
+* [major] Move JToken-specific members from `JsonUtility` to `JTokenUtility`.
 
 ## Released